requirements.txt
asgiref==3.9.1
attrs==25.3.0
boto3==1.39.4
botocore==1.39.4
cffi==1.17.1
cfgv==3.4.0
cryptography==45.0.5
distlib==0.3.9
Django==5.2.4
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.0
drf-spectacular==0.28.0
filelock==3.18.0
flake8==7.3.0
gunicorn==23.0.0
identify==2.6.12
inflection==0.5.1
iniconfig==2.1.0
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
nodeenv==1.9.1
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
Pygments==2.19.2
PyMySQL==1.1.1
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.31.2
